Adding the md for Jaboca berry
Very rough atm, all links point the homepage, and there's a lot of things missing (and I don't know if the table of content I made will work), but I want to take a first look at it on the repo itself.
Missing FlowChart
Missing colour in the mutation table, probably needs html to work.