Wandering Pokémon (#71)
* Wandering Pokémon

Adds a page with information about Wandering Pokémon. It lists the base wanderers, color wanderers, and specific wanderers.

To help people check which berry is which color, added the "Color" column to the main Berry page. No need to add it to individual berry pages since those already had the data.

Also added a link to this page to the Farm page.

* Typeahead

Typeahead, please. Also fixed a typo.

* Filter

Thank you Jessica for the filter simplification.

---------

Co-authored-by: Danial <RedSparr0w@users.noreply.github.com>