Fixed default BF calculator (#84)
* Fixed default BF calculator

The BF calculator was loading stage 1k for some reason instead of stage 100 and thus showing wrong information. Now it loads... stage 1k anyway but it shows the correct stats.

* Changed fixed values for last milestone

The way the BF calculator works, it stores the last milestone so, with this, the default will now be whichever is the last milestone.