NPCs (#193)
* NPCs

Initial draft for NPCs page. Doesn't have a clickable table nor individual pages because, for now, it wouldn't be simple.

* Link To Town

Since linking to an individual page is too complex for now, I decided to make the page link to Town instead. That way people can click it, get the details about the location, and check the NPC in-game.

---------

Co-authored-by: Danial <RedSparr0w@users.noreply.github.com>